refactor: enhance docs and add comments
