+++ .github/workflows/test.yml
@@ -1,6 +1,8 @@
+# This is a workflow for testing regression
 name: Test
 on:
   workflow_dispatch:
+  # only run when pull request contains typescript file changes in src folder
   pull_request:
     paths:
       - src/**/*.ts
