+++ .gitignore
@@ -12,6 +12,8 @@ dist
 dist-ssr
 *.local
 
+test-results.json
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
