+++ .storybook/preview.ts
@@ -1,6 +1,5 @@
 import type { Preview } from "@storybook/react";
 
-
 const preview: Preview = {
   parameters: {
     controls: {
