+++ package.json
@@ -13,9 +13,10 @@
     "storybook": "storybook dev -p 6006",
     "build:storybook": "storybook build",
     "test:storybook": "test-storybook --json --outputFile results.json",
-    "test:playwright": "npx playwright install && npx playwright test --ui",
     "test": "vitest run --browser.headless",
-    "test:coverage": "vitest --coverage"
+    "test:coverage": "vitest --coverage",
+    "test:ui": "vitest",
+    "test:visual": "npx playwright install && npx playwright test --ui"
   },
   "dependencies": {
     "@arcgis/core": "^4.32.3",
@@ -55,7 +56,6 @@
     "eslint-plugin-react-refresh": "^0.4.16",
     "eslint-plugin-storybook": "^0.11.2",
     "globals": "^15.14.0",
-    "happy-dom": "^17.1.8",
     "jsdom": "^26.0.0",
     "nodemon": "^3.1.9",
     "playwright": "^1.50.1",
