+++ playwright.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
-  // Look for test files in the "tests" directory, relative to this configuration file.
-  testDir: "__tests__",
+  // Look for test files in the "__visualregression__" directory, relative to this configuration file.
+  testDir: "__visualregression__",
 
   // Run all tests in parallel.
   fullyParallel: true,
@@ -14,11 +14,11 @@ export default defineConfig({
   workers: process.env.CI ? 1 : undefined,
 
   // Reporter to use
-  reporter: "html",
+  reporter: "json",
 
   use: {
     // Base URL to use in actions like `await page.goto('/')`.
-    baseURL: "http://localhost:3001",
+    baseURL: "http://localhost:3000",
 
     // Collect trace when retrying the failed test.
     trace: "on-first-retry",
@@ -43,7 +43,7 @@ export default defineConfig({
   // Run your local dev server before starting the tests.
   webServer: {
     command: "npm run dev",
-    url: "http://localhost:3001",
+    url: "http://localhost:3000",
     reuseExistingServer: !process.env.CI,
   },
 });
