+++ test-results.json
@@ -1,103 +1,1 @@
-{
-  "numTotalTestSuites": 6,
-  "numPassedTestSuites": 6,
-  "numFailedTestSuites": 0,
-  "numPendingTestSuites": 0,
-  "numTotalTests": 5,
-  "numPassedTests": 5,
-  "numFailedTests": 0,
-  "numPendingTests": 0,
-  "numTodoTests": 0,
-  "snapshot": {
-    "added": 0,
-    "failure": false,
-    "filesAdded": 0,
-    "filesRemoved": 0,
-    "filesRemovedList": [],
-    "filesUnmatched": 0,
-    "filesUpdated": 0,
-    "matched": 0,
-    "total": 0,
-    "unchecked": 0,
-    "uncheckedKeysByFile": [],
-    "unmatched": 0,
-    "updated": 0,
-    "didUpdate": false
-  },
-  "startTime": 1741115888543,
-  "success": true,
-  "testResults": [
-    {
-      "assertionResults": [
-        {
-          "ancestorTitles": ["fetchData functions"],
-          "fullName": "fetchData functions loadObservations should fetch and return data",
-          "status": "passed",
-          "title": "loadObservations should fetch and return data",
-          "duration": 2.599999999976717,
-          "failureMessages": [],
-          "meta": {}
-        },
-        {
-          "ancestorTitles": ["fetchData functions"],
-          "fullName": "fetchData functions saveObservation should send a POST request and return response",
-          "status": "passed",
-          "title": "saveObservation should send a POST request and return response",
-          "duration": 0.7000000000698492,
-          "failureMessages": [],
-          "meta": {}
-        }
-      ],
-      "startTime": 1741115951705,
-      "endTime": 1741115951708.7,
-      "status": "passed",
-      "message": "",
-      "name": "C:/Users/danie/Desktop/projects/testing-for-success-arcgis-maps-sdk-for-javascript/src/api/__tests__/fetchData.test.ts"
-    },
-    {
-      "assertionResults": [
-        {
-          "ancestorTitles": ["DataEntry"],
-          "fullName": "DataEntry renders DataEntry, handles properties, and displays latitude and longitude",
-          "status": "passed",
-          "title": "renders DataEntry, handles properties, and displays latitude and longitude",
-          "duration": 1058.3999999999069,
-          "failureMessages": [],
-          "meta": {}
-        }
-      ],
-      "startTime": 1741115951875,
-      "endTime": 1741115952933.4,
-      "status": "passed",
-      "message": "",
-      "name": "C:/Users/danie/Desktop/projects/testing-for-success-arcgis-maps-sdk-for-javascript/src/components/__tests__/DataEntry.test.tsx"
-    },
-    {
-      "assertionResults": [
-        {
-          "ancestorTitles": ["MapContainer"],
-          "fullName": "MapContainer webgl enabled",
-          "status": "passed",
-          "title": "webgl enabled",
-          "duration": 21,
-          "failureMessages": [],
-          "meta": {}
-        },
-        {
-          "ancestorTitles": ["MapContainer"],
-          "fullName": "MapContainer renders and trigger onMapLoad when map is ready",
-          "status": "passed",
-          "title": "renders and trigger onMapLoad when map is ready",
-          "duration": 2147.20000000007,
-          "failureMessages": [],
-          "meta": {}
-        }
-      ],
-      "startTime": 1741115953128,
-      "endTime": 1741115955296.2,
-      "status": "passed",
-      "message": "",
-      "name": "C:/Users/danie/Desktop/projects/testing-for-success-arcgis-maps-sdk-for-javascript/src/components/__tests__/MapContainer.test.tsx"
-    }
-  ]
-}
+{"numTotalTestSuites":2,"numPassedTestSuites":2,"numFailedTestSuites":0,"numPendingTestSuites":0,"numTotalTests":2,"numPassedTests":1,"numFailedTests":0,"numPendingTests":1,"numTodoTests":0,"snapshot":{"added":0,"failure":false,"filesAdded":0,"filesRemoved":0,"filesRemovedList":[],"filesUnmatched":0,"filesUpdated":0,"matched":0,"total":0,"unchecked":0,"uncheckedKeysByFile":[],"unmatched":0,"updated":0,"didUpdate":false},"startTime":1741134987076,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":["MapContainer"],"fullName":"MapContainer webgl enabled","status":"skipped","title":"webgl enabled","failureMessages":[],"location":{"line":44,"column":3},"meta":{}},{"ancestorTitles":["MapContainer"],"fullName":"MapContainer renders and trigger onMapLoad when map is ready","status":"passed","title":"renders and trigger onMapLoad when map is ready","duration":254.30000000074506,"failureMessages":[],"location":{"line":50,"column":3},"meta":{}}],"startTime":1741135022591,"endTime":1741135022845.3,"status":"passed","message":"","name":"C:/Users/danie/Desktop/projects/testing-for-success-arcgis-maps-sdk-for-javascript/src/components/__tests__/MapContainer.test.tsx"}]}
